Update lamda code to latest on aws

+++ src/dynamoDBLambda/dynamodb_api.js
@@ -8,8 +8,8 @@ const dynamo = new doc.DynamoDB();
  * Demonstrates a simple HTTP endpoint using API Gateway.
  *
  * I use POST with a request body for all methods.
- * myMethod is used to define the mongoDB method.
- * I like to do this for security reasons because query strings
+ * myMethod is used to define the method.
+ * I like to do this for security reasons because the query strings
  * in URL can be logged on the server.
  *
  * The following JSON object is an example for dynamo.putItem.
@@ -27,7 +27,7 @@ const dynamo = new doc.DynamoDB();
  *          myMethod: 'putItem',
  *        }
  *
- * I also had to add to these items to the headers for CORS
+ * I also had to add to get is to work with CORS
  * 'Access-Control-Allow-Origin': '*',
  * 'Access-Control-Allow-Methods': "POST",
  *
@@ -44,19 +44,34 @@ exports.handler = (event, context, callback) => {
         'Access-Control-Allow-Methods': 'POST',
       },
     });
-
+  const lowestTeamNumber = 10;
   switch (event.httpMethod) {
     case 'POST':
       var myEventBody = JSON.parse(event.body);
       switch (myEventBody.myMethod) {
         case 'deleteItem':
-          dynamo.deleteItem(myEventBody.myBody, done);
+          if (myEventBody.myBody.Key.team_id <= lowestTeamNumber) {
+            done(new Error(`Unsupported method "${event.httpMethod}"`));
+          } else {
+            dynamo.deleteItem(myEventBody.myBody, done);
+          }
           break;
         case 'putItem':
-          dynamo.putItem(myEventBody.myBody, done);
+          if (myEventBody.myBody.Item.team_id <= lowestTeamNumber) {
+            done(new Error(`Unsupported method "${event.httpMethod}"`));
+          } else {
+            dynamo.putItem(myEventBody.myBody, done);
+          }
           break;
         case 'updateItem':
-          dynamo.updateItem(myEventBody.myBody, done);
+          if (myEventBody.myBody.Item.team_id <= lowestTeamNumber) {
+            done(new Error(`Unsupported method "${event.httpMethod}"`));
+          } else {
+            dynamo.updateItem(myEventBody.myBody, done);
+          }
+          break;
+        case 'getItem':
+          dynamo.getItem(myEventBody.myBody, done);
           break;
         case 'scan':
           dynamo.scan(myEventBody.myBody, done);
